FEATURE: Extract node update to separate method
As it can happen that you may not use the PublishingService in the Backend, you probably don’t have a BackendUser. So this change extracts the update of the node to an extra method that just gets a string for the current publisher.

We also define the maximum items via setting.